requirements.txt
numpy<2.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn
matplotlib
opencv-python
Pillow
pycocotools
decord
einops
open_clip_torch
faiss-cpu  
timm
transformers
spacy
openai
huggingface_hub
tqdm
gdown
mega.py
nest_asyncio
ftfy
regex